+++ contract/transfers/programs/transfers/src/instructions/mod.rs
@@ -1,0 +1,3 @@
+pub mod transfer;
+
+pub use transfer::*;
